requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astroid==2.15.8
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
bandit==1.7.5
bardapi==0.1.38
beautifulsoup4==4.12.2
black==22.12.0
blinker==1.7.0
blis==0.7.11
browser-cookie3==0.19.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudscraper==1.2.71
colorama==0.4.6
confection==0.1.4
coverage==6.5.0
coverage-badge==1.1.0
cryptography==41.0.7
cymem==2.0.8
darglint==1.8.1
dataclasses-json==0.6.3
deckar01-ratelimit==3.0.2
deep-translator==1.11.4
deepl==1.16.1
dill==0.3.7
diskcache==5.6.3
distlib==0.3.7
Django==4.2.7
django-postgres-extra==2.0.8
django-redis==5.4.0
django-rq==2.9.0
dnspython==2.4.2
dparse==0.6.3
elastic-transport==8.10.0
elasticsearch==8.11.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
english-words==2.0.1
et-xmlfile==1.1.0
filelock==3.13.1
flake8==6.1.0
frozenlist==1.4.0
fsspec==2023.12.0
ftfy==6.1.3
futures==3.0.5
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.14.0
google-auth==2.24.0
google-cloud-aiplatform==1.36.4
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.13.0
google-cloud-translate==3.12.1
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
goslate==1.5.4
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
hyperframe==6.0.1
identify==2.5.32
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.345
langchain-core==0.0.9
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.69
lazy-object-proxy==1.9.0
libretranslatepy==2.1.1
llama_cpp_python==0.2.20
locate==1.1.1
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
mdurl==0.1.2
meteor==0.1.0
meteor-ejson==1.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
mypy==1.7.1
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.24.4
openai==0.27.10
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
pbr==6.0.0
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==4.0.0
pluggy==1.3.0
pre-commit==2.21.0
preshed==3.0.9
proto-plus==1.22.3
protobuf==4.25.1
py==1.11.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydocstyle==6.3.0
pyee==11.1.0
pyflakes==3.1.0
Pygments==2.17.2
pylint==2.17.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyparsing==3.1.1
pypdf==3.17.1
PyPDF2==3.0.1
pyphen==0.14.0
pyspellchecker==0.7.2
pytest==7.4.3
pytest-cov==3.0.0
pytest-django==4.7.0
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-snapshot==0.9.0
python-dateutil==2.8.2
python-ddp==0.1.5
python-meteor==0.1.6
pytz==2023.3.post1
pyupgrade==2.38.4
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rich==13.7.0
rouge==1.0.1
rpds-py==0.13.2
rq==1.15.1
rsa==4.9
safetensors==0.4.1
safety==1.10.3
scikit-learn==1.3.2
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.38.0
shapely==2.0.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
snapshot==0.1.2
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
spacy==3.7.2
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.3
SQLAlchemy==2.0.23
sqlparse==0.4.4
srsly==2.4.8
stevedore==5.1.0
streamlit==1.29.0
structlog==23.2.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
textmetrics==0.0.2
textstat==0.7.3
thinc==8.2.1
threadpoolctl==3.2.0
timeout-decorator==0.5.0
tldextract==5.1.1
tokenize-rt==4.2.1
tokenizers==0.15.0
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
transformers==4.35.2
translate==3.6.1
typer==0.9.0
types-pyOpenSSL==23.3.0.0
types-redis==4.6.0.11
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
vertexai==0.0.1
virtualenv==20.25.0
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
wikipedia==1.4.0
wordfreq==3.1.1
wordhoard==1.5.4
wrapt==1.16.0
ws4py==0.5.1
yarl==1.9.3
zipp==3.17.0
